+++ branch-alpha/sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2026-08-11T07:33:05+0000
+// 2026-08-12T01:42:46+0000
 (function () {
 "use strict";